+++ restaurant/src/front_of_house.rs
@@ -1,0 +1,9 @@
+pub mod hosting;
+
+mod serving {
+    fn take_order() {}
+
+    fn serve_order() {}
+
+    fn take_payment() {}
+}
