+++ restaurant/src/front_of_house/hosting.rs
@@ -1,0 +1,5 @@
+pub fn add_to_waitlist() {
+    println!("add_to_waitlist");
+}
+
+fn seat_at_table() {}
