+++ restaurant/src/lib.rs
@@ -1,28 +1,31 @@
 // 関数に対してuseを使う場合は関数名を含めてしまうと、
 // 関数がどこから持ち込まれたかわからなくなるので関数名は含めないのが慣例
 // useと絶対パス
-use crate::front_of_house::hosting;
+// use crate::front_of_house::hosting;
 // useと相対パス
-use self::front_of_house::hosting;
+// use self::front_of_house::hosting;
+// pub use crate::front_of_house::hosting;
+
+mod front_of_house;
 pub use crate::front_of_house::hosting;
 
-mod front_of_house {
-    pub mod hosting {
-        pub fn add_to_waitlist() {
-            println!("add_to_waitlist");
-        }
-
-        fn seat_at_table() {}
-    }
-
-    mod serving {
-        fn take_order() {}
-
-        fn serve_order() {}
-
-        fn take_payment() {}
-    }
-}
+// mod front_of_house {
+//     pub mod hosting {
+//         pub fn add_to_waitlist() {
+//             println!("add_to_waitlist");
+//         }
+//
+//         fn seat_at_table() {}
+//     }
+//
+//     mod serving {
+//         fn take_order() {}
+//
+//         fn serve_order() {}
+//
+//         fn take_payment() {}
+//     }
+// }
 
 fn serve_order() {
     println!("serve_order");
@@ -64,42 +67,46 @@ mod back_of_house {
 }
 
 pub fn eat_at_restaurant() {
-    // パスをpubキーワードで公開する
+    // // パスをpubキーワードで公開する
+    // {
+    //     // 絶対パス
+    //     crate::front_of_house::hosting::add_to_waitlist();
+    //
+    //     // 相対パス
+    //     front_of_house::hosting::add_to_waitlist();
+    //
+    //     // useキーワード使用
+    //     hosting::add_to_waitlist();
+    // }
+    //
+    // // 相対パスをsuperで始める
+    // {
+    //     back_of_house::fix_incorrect_order();
+    // }
+    //
+    // // 構造体とenumを公開する
+    // {
+    //     // 夏 (summer) にライ麦 (Rye) パン付き朝食を注文
+    //     let mut meal = back_of_house::Breakfast::summer("Rye");
+    //     // やっぱり別のパンにする
+    //     meal.toast = String::from("Wheat");
+    //     println!("I'd like {} toast please", meal.toast);
+    //
+    //     // seasonal_fruitは公開されていないので、下の行はコンパイルできない。
+    //     // meal.seasonal_fruit = String::from("blueberries");
+    //
+    //     // seasonal_fruitは公開されていないので、以下のようなインスタンス生成は出来ない
+    //     // そのためインスタンス生成用の関連関数summerを用意している
+    //     // let mut meal = crate::back_of_house::Breakfast {
+    //     //     toast: String::from("hoge"),
+    //     //     seasonal_fruit: String::from("fuga")
+    //     // };
+    //
+    //     let order1 = back_of_house::Appetizer::Soup;
+    //     let order2 = back_of_house::Appetizer::Salad;
+    // }
+
     {
-        // 絶対パス
-        crate::front_of_house::hosting::add_to_waitlist();
-
-        // 相対パス
-        front_of_house::hosting::add_to_waitlist();
-
-        // useキーワード使用
         hosting::add_to_waitlist();
-    }
-
-    // 相対パスをsuperで始める
-    {
-        back_of_house::fix_incorrect_order();
-    }
-
-    // 構造体とenumを公開する
-    {
-        // 夏 (summer) にライ麦 (Rye) パン付き朝食を注文
-        let mut meal = back_of_house::Breakfast::summer("Rye");
-        // やっぱり別のパンにする
-        meal.toast = String::from("Wheat");
-        println!("I'd like {} toast please", meal.toast);
-
-        // seasonal_fruitは公開されていないので、下の行はコンパイルできない。
-        // meal.seasonal_fruit = String::from("blueberries");
-
-        // seasonal_fruitは公開されていないので、以下のようなインスタンス生成は出来ない
-        // そのためインスタンス生成用の関連関数summerを用意している
-        // let mut meal = crate::back_of_house::Breakfast {
-        //     toast: String::from("hoge"),
-        //     seasonal_fruit: String::from("fuga")
-        // };
-
-        let order1 = back_of_house::Appetizer::Soup;
-        let order2 = back_of_house::Appetizer::Salad;
     }
 }
